First fee calculations work for closing a short position
